backend/Ewaste_backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.7.0
djangorestframework==3.17.1
PyJWT==2.10.1
python-dotenv==1.2.2
sqlparse==0.5.3
tzdata==2025.3
